flask_app/requirements.txt
aniso8601==3.0.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-WTF==1.0.0
google-api-core==2.6.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pytz==2018.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.11.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==0.14.1
WTForms==3.0.1
